Moved all classes to presentation.ui.security package
