+++ owncloudApp/src/main/AndroidManifest.xml
@@ -225,7 +225,7 @@
             android:configChanges="keyboard|keyboardHidden|orientation|screenSize|screenLayout|smallestScreenSize|uiMode"
             android:theme="@style/Theme.ownCloud.Fullscreen" />
         <activity android:name=".presentation.ui.security.PatternActivity" />
-        <activity android:name=".ui.activity.BiometricActivity" />
+        <activity android:name=".presentation.ui.security.BiometricActivity" />
         <activity
             android:name=".presentation.ui.authentication.LoginActivity"
             android:exported="true"
