+++ owncloudApp/src/main/java/com/owncloud/android/presentation/ui/security/PatternManager.kt
@@ -26,7 +26,6 @@ import android.content.Intent
 import android.os.Build
 import android.os.PowerManager
 import com.owncloud.android.MainApp.Companion.appContext
-import com.owncloud.android.authentication.BiometricManager
 import com.owncloud.android.data.preferences.datasources.implementation.SharedPreferencesProviderImpl
 
 object PatternManager {
